Fleshed out PlayerModel

+++ Sources/Domain/Video.swift
@@ -14,7 +14,7 @@ struct Video: Sendable, Codable, Hashable, Equatable {
     let duration: String   // ISO-8601 string as returned by YouTube API (e.g. "PT3M42S")
     let url: String
     let publishedAt: Date
-    var startAt: Double = 0          // seconds offset where playback should start (default 0)
+    var startAt: Int = 0          // seconds offset where playback should start (default 0)
 }
 
 extension Video {
